Add buggy button to leave the room
This button makes the user leave the room, but there are a lot of bugs
that occur after using this button. At this point, the codebase is so
large that we need to use TDD to add new features. The next few commits
will hopefully implement TDD.

After we have a large test suite, we can find the cause of the bugs and
remove them. After that, we can merge this back into the main branch.
Until then, we cannot safely deploy the codebase as present in this
commit.