Refactor footer code and add test
I wanted to add a test for the simplest component first and then add
tests for the more complex components. Footer is just a static
component, so it is somewhat easy to test. However, I do not know how to
test the images themselves, so I just have the test check to see if
there are elements with the appropriate alt test.

The next steps will be to add tests to the other components. Also, the
tests all pass on the footer.