Manage tests using Jest
Prior to this commit, we already used testing-library to run tests, but
it was a bit messy to run them ourselves. This commit adds Jest to our
codebase. Jest manages all our tests, and it integrates into VSCode.

I set up Jest in the root directory, so it can test both our frontend
and our backend. Right now, only the frontend has tests, but it will be
easier this way to add tests to the backend. Also, I added Typescript
support to Jest. I also moved the tsconfig.json file to the root
directory because Jest needs access to it there. This also means that
the backend will support TS.